+++ CHANGELOG.md
@@ -8,6 +8,7 @@ From v1.0.0 onwards, this project adheres to [Semantic Versioning](https://semve
 
 ## Master (Unreleased)
 
+- Add support for nested test classes
 - Removes `_snapshot_subdirectory_name` from `SnapshotFossilizer`
 
 ## v0.1.0